refactor: index.js에 있던 함수 model/model.js로 이동
index.js에 어울리지 않았던 회원가입함수들을 분리하여 model/model.js로
이동
- 이후 클래스로 바꾸고, 데이터 저장 형식도 바꿀 예정
