+++ routes/index.js
@@ -24,48 +24,6 @@ const listPosts = () => async (req, res, next) => {
     }
 }
 
-const isFileExit = (filePath) => {
-    return new Promise((resolve, reject) => {
-        fs.access(filePath, (err) => {
-            if (err) reject('notExit');
-            resolve('exit');
-        })
-    });
-}
-
-const signUp = () => async (req, res, next) => {
-    const { id } = req.body;
-    try {
-        const IDexit = await isFileExit(`users/${id}.json`);
-        res.end(IDexit);
-    } catch (notExit) {
-        res.end(notExit);
-    }
-}
-
-const createAccount = (filePath, data) => {
-    return new Promise((resolve, reject) => {
-        fs.writeFile(filePath, data, (err) => {
-            if (err) reject('fail');
-            resolve('Created account. Enjoy Todo List!');
-        })
-    });
-}
-
-const createID = () => async (req, res, next) => {
-    const { id, pwd } = req.body;
-    const filePath = `users/${id}.json`
-    const data = `{ "id": "${id}", "pwd": "${pwd}"}`
-    try {
-        const note = await createAccount(filePath, data);
-        res.end(note);
-    } catch (fail) {
-        res.end(fail);
-    }
-}
-
 module.exports = {
-    listPosts,
-    signUp,
-    createID,
+    listPosts
 }
