+++ users/jake7.json
@@ -1,0 +1,1 @@
+{ "id": "jake7", "pwd": "jake1234"}
